[FE] feat: 레이블 페이지 구현 (#93)
