+++ fe/src/components/labelListPage/Body.tsx
@@ -1,14 +1,16 @@
 import { useTheme } from '@emotion/react';
 import { Box } from '@components/common/box/Box';
 import { LabelItem } from './LabelItem';
-import { TableContainer } from '@components/common/Table/TableContainer';
+
 import { TableHeader } from '@components/common/Table/TableHeader';
+import { LabelEditTable } from './LabelEditTable';
 
 type Props = {
   labelCount: IssuePageData['labelCount'];
   labelList: Label[];
   isAddTableOpen?: boolean;
-  onAddTableClose?: () => void;
+  onAddTableClose: () => void;
+  fetchLabelList: () => Promise<void>;
 };
 
 export const Body: React.FC<Props> = ({
@@ -16,6 +18,7 @@ export const Body: React.FC<Props> = ({
   labelList,
   isAddTableOpen,
   onAddTableClose,
+  fetchLabelList,
 }) => {
   const theme = useTheme() as any;
 
@@ -29,11 +32,11 @@ export const Body: React.FC<Props> = ({
       }}
     >
       {isAddTableOpen && (
-        <TableContainer
-          tableVariant="label"
+        <LabelEditTable
+          header={<TableHeader title="새로운 레이블 추가" />}
           typeVariant="add"
           onAddTableClose={onAddTableClose}
-          header={<TableHeader title="새로운 레이블 추가" />}
+          fetchLabelList={fetchLabelList}
         />
       )}
 
@@ -53,10 +56,8 @@ export const Body: React.FC<Props> = ({
         {labelList.map((label) => (
           <LabelItem
             key={label.id}
-            name={label.name}
-            textColor={label.textColor}
-            backgroundColor={label.backgroundColor}
-            description={label.description}
+            label={label}
+            fetchLabelList={fetchLabelList}
           />
         ))}
       </Box>
