+++ fe/src/components/labelListPage/LabelItem.tsx
@@ -5,22 +5,15 @@ import { InformationTag } from '@components/common/InformationTag';
 import { Button } from '@components/common/Button';
 import { ReactComponent as Edit } from '@assets/icons/edit.svg';
 import { ReactComponent as Trash } from '@assets/icons/trash.svg';
-import { TableContainer } from '@components/common/Table/TableContainer';
 import { TableHeader } from '@components/common/Table/TableHeader';
+import { LabelEditTable } from './LabelEditTable';
 
 type Props = {
-  name: string;
-  textColor: ThemeType;
-  backgroundColor: string;
-  description: string;
+  label: Label;
+  fetchLabelList: () => Promise<void>;
 };
 
-export const LabelItem: React.FC<Props> = ({
-  name,
-  textColor,
-  backgroundColor,
-  description,
-}) => {
+export const LabelItem: React.FC<Props> = ({ label, fetchLabelList }) => {
   const theme = useTheme() as any;
   const [isEditing, setIsEditing] = useState(false);
 
@@ -32,10 +25,25 @@ export const LabelItem: React.FC<Props> = ({
     setIsEditing(false);
   };
 
+  const onDeleteLabel = async () => {
+    const response = await fetch(
+      `${import.meta.env.VITE_APP_BASE_URL}/labels/${label.id}`,
+      {
+        method: 'DELETE',
+      },
+    );
+    if (response.ok) {
+      if (response.headers.get('content-type')?.includes('application/json')) {
+        const data = await response.json();
+        console.log(data);
+      }
+      await fetchLabelList();
+    }
+  };
+
   return (
     <li
       css={{
-        // height: '96px',
         width: '100%',
         display: 'flex',
         alignItems: 'center',
@@ -46,12 +54,13 @@ export const LabelItem: React.FC<Props> = ({
       }}
     >
       {isEditing ? (
-        <TableContainer
-          tableVariant="label"
+        <LabelEditTable
+          header={<TableHeader title="레이블 편집" />}
+          fetchLabelList={fetchLabelList}
+          label={label}
           typeVariant="edit"
           onAddTableClose={onEditLabelClose}
-          header={<TableHeader title="레이블 편집" />}
-        ></TableContainer>
+        />
       ) : (
         <>
           <div
@@ -64,10 +73,10 @@ export const LabelItem: React.FC<Props> = ({
             <InformationTag
               size="S"
               typeVariant="filled"
-              fillColor={backgroundColor}
-              textColor={textColor}
+              fillColor={label.backgroundColor}
+              textColor={label.textColor}
             >
-              {name}
+              {label.name}
             </InformationTag>
           </div>
 
@@ -78,7 +87,7 @@ export const LabelItem: React.FC<Props> = ({
               font: theme.fonts.displayMedium16,
             }}
           >
-            {description}
+            {label.description}
           </span>
           <div
             css={{
@@ -107,6 +116,7 @@ export const LabelItem: React.FC<Props> = ({
                 padding: '0',
                 color: theme.danger.text.default,
               }}
+              onClick={onDeleteLabel}
             >
               <Trash stroke={theme.danger.text.default} />
               삭제
