+++ fe/src/utils/api.ts
@@ -2,7 +2,6 @@ export const fetchData = async (path: string) => {
   try {
     const response = await fetch(
       `${import.meta.env.VITE_APP_BASE_URL}/${path}`,
-      // `${path}`,
       {
         method: 'GET',
       },
